+++ src/main/kotlin/ar/edu/unq/peluqueriayabackend/service/GenericService.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 interface GenericService<T> {
 
-    fun get(id:Int): Optional<T>
+    fun get(id:Long): Optional<T>
     fun getAll():Collection<T>
     fun save(t:T):T
     fun update(t:T):T
